Final Checkin for Popular Movies Phase 1

+++ app/src/main/java/com/ronsparks/popularmovies/data/MovieItem.java
@@ -26,6 +26,6 @@ public class MovieItem {
     public Long movieId;
     public Double popularity;
     public Integer voteCount;
-    public Long voteAverage;
+    public Double voteAverage;
     //endregion
 }
